pyproject.toml
[project]
name = "v2-one"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.18.4",
    "httpx>=0.28.1",
    "ollama>=0.6.1",
    "psycopg2-binary>=2.9.11",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
]

[project.scripts]
v2-one = "v2_one.main:main"

[dependency-groups]
dev = [
    "black>=22.3.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
